+++ tests/__init__.py
@@ -18,6 +18,8 @@ import numpy as np
 TEST_ROOT = os.path.dirname(__file__)
 PROJECT_ROOT = os.path.dirname(TEST_ROOT)
 TEMP_PATH = os.path.join(PROJECT_ROOT, 'test_temp')
+DATASETS_PATH = os.path.join(PROJECT_ROOT, 'Datasets')
+LEGACY_PATH = os.path.join(PROJECT_ROOT, 'legacy')
 
 # todo: this setting `PYTHONPATH` may not be used by other evns like Conda for import packages
 if PROJECT_ROOT not in os.getenv('PYTHONPATH', ""):